refactor(tasks): run manager implemented - it handles all logic regarding running the job
This improves clarity of the main task handler service. All run state logic should be eventually moved here.
